+++ app/vlselect/logsql/logsql.go
@@ -12,7 +12,6 @@ import (
 	"strconv"
 	"strings"
 	"sync"
-	"sync/atomic"
 	"time"
 
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/atomicutil"
@@ -931,22 +930,8 @@ func ProcessQueryRequest(ctx context.Context, w http.ResponseWriter, r *http.Req
 
 	if limit > 0 {
 		if q.CanReturnLastNResults() {
-			rows, err := getLastNQueryResults(ctx, tenantIDs, q, limit)
-			if err != nil {
-				httpserver.Errorf(w, r, "%s", err)
-				return
-			}
-			bw := bwShards.Get(0)
-			for i := range rows {
-				bw.buf = logstorage.MarshalFieldsToJSON(bw.buf, rows[i].fields)
-				bw.buf = append(bw.buf, '\n')
-				if len(bw.buf) > 16*1024 {
-					bw.FlushIgnoreErrors()
-				}
-			}
-			return
+			q.AddPipeSortByTimeDesc()
 		}
-
 		q.AddPipeLimit(uint64(limit))
 	}
 
@@ -1000,137 +985,6 @@ func (bw *bufferedWriter) Write(p []byte) (int, error) {
 func (bw *bufferedWriter) FlushIgnoreErrors() {
 	_, _ = bw.sw.Write(bw.buf)
 	bw.buf = bw.buf[:0]
-}
-
-func getLastNQueryResults(ctx context.Context, tenantIDs []logstorage.TenantID, q *logstorage.Query, limit int) ([]logRow, error) {
-	limitUpper := 2 * limit
-	q.AddPipeLimit(uint64(limitUpper))
-
-	rows, err := getQueryResultsWithLimit(ctx, tenantIDs, q, limitUpper)
-	if err != nil {
-		return nil, err
-	}
-	if len(rows) < limitUpper {
-		// Fast path - the requested time range contains up to limitUpper rows.
-		rows = getLastNRows(rows, limit)
-		return rows, nil
-	}
-
-	// Slow path - adjust time range for selecting up to limitUpper rows
-	start, end := q.GetFilterTimeRange()
-	d := end/2 - start/2
-	start += d
-
-	qOrig := q
-	for {
-		timestamp := qOrig.GetTimestamp()
-		q = qOrig.CloneWithTimeFilter(timestamp, start, end)
-		rows, err := getQueryResultsWithLimit(ctx, tenantIDs, q, limitUpper)
-		if err != nil {
-			return nil, err
-		}
-
-		if d == 0 || start >= end {
-			// The [start ... end] time range equals one nanosecond.
-			// Just return up to limit rows.
-			if len(rows) > limit {
-				rows = rows[:limit]
-			}
-			return rows, nil
-		}
-
-		dLastBit := d & 1
-		d /= 2
-
-		if len(rows) >= limitUpper {
-			// The number of found rows on the [start ... end] time range exceeds limitUpper,
-			// so reduce the time range to [start+d ... end].
-			start += d
-			continue
-		}
-		if len(rows) >= limit {
-			// The number of found rows is in the range [limit ... limitUpper).
-			// This means that found rows contains the needed limit rows with the biggest timestamps.
-			rows = getLastNRows(rows, limit)
-			return rows, nil
-		}
-
-		// The number of found rows on [start ... end] time range is below the limit.
-		// This means the time range doesn't cover the needed logs, so it must be extended.
-
-		if len(rows) == 0 {
-			// The [start ... end] time range doesn't contain any rows, so change it to [start-d ... start).
-			end = start - 1
-			start -= d + dLastBit
-			continue
-		}
-
-		// The number of found rows on [start ... end] time range is bigger than 0 but smaller than limit.
-		// Increase the time range to [start-d ... end].
-		start -= d + dLastBit
-	}
-}
-
-func getLastNRows(rows []logRow, limit int) []logRow {
-	sortLogRows(rows)
-	if len(rows) > limit {
-		rows = rows[len(rows)-limit:]
-	}
-	return rows
-}
-
-func getQueryResultsWithLimit(ctx context.Context, tenantIDs []logstorage.TenantID, q *logstorage.Query, limit int) ([]logRow, error) {
-	ctxWithCancel, cancel := context.WithCancel(ctx)
-	defer cancel()
-
-	var missingTimeColumn atomic.Bool
-	var rows []logRow
-	var rowsLock sync.Mutex
-	writeBlock := func(_ uint, db *logstorage.DataBlock) {
-		if missingTimeColumn.Load() {
-			return
-		}
-
-		columns := db.Columns
-		clonedColumnNames := make([]string, len(columns))
-		for i, c := range columns {
-			clonedColumnNames[i] = strings.Clone(c.Name)
-		}
-
-		timestamps, ok := db.GetTimestamps(nil)
-		if !ok {
-			missingTimeColumn.Store(true)
-			cancel()
-			return
-		}
-
-		for i, timestamp := range timestamps {
-			fields := make([]logstorage.Field, len(columns))
-			for j := range columns {
-				f := &fields[j]
-				f.Name = clonedColumnNames[j]
-				f.Value = strings.Clone(columns[j].Values[i])
-			}
-
-			rowsLock.Lock()
-			rows = append(rows, logRow{
-				timestamp: timestamp,
-				fields:    fields,
-			})
-			rowsLock.Unlock()
-		}
-
-		if len(rows) >= limit {
-			cancel()
-		}
-	}
-	err := vlstorage.RunQuery(ctxWithCancel, tenantIDs, q, writeBlock)
-
-	if missingTimeColumn.Load() {
-		return nil, fmt.Errorf("missing _time column in the result for the query [%s]", q)
-	}
-
-	return rows, err
 }
 
 func parseCommonArgs(r *http.Request) (*logstorage.Query, []logstorage.TenantID, error) {